requirement.txt
tensorflow-gpu==2.4.0
matplotlib==3.5.1